+++ tests/test_stac_api_io.py
@@ -1,4 +1,5 @@
 import pytest
+from pystac_client.conformance import ConformanceClasses
 
 from pystac_client.exceptions import APIError
 from pystac_client.stac_api_io import StacApiIO
@@ -37,3 +38,24 @@ class TestSTAC_IOOverride:
         response = stac_api_io.read_text(str(test_file))
 
         assert response == 'Hi there!'
+
+    def test_assert_conforms_to(self):
+        nonconformant = StacApiIO(conformance=[])
+
+        with pytest.raises(NotImplementedError):
+            nonconformant.assert_conforms_to(ConformanceClasses.CORE)
+
+        conformant_io = StacApiIO(conformance=["https://api.stacspec.org/v1.0.0-beta.1/core"])
+
+        # Check that this does not raise an exception
+        conformant_io.assert_conforms_to(ConformanceClasses.CORE)
+
+    def test_conforms_to(self):
+        nonconformant = StacApiIO(conformance=[])
+
+        assert not nonconformant.conforms_to(ConformanceClasses.CORE)
+
+        conformant_io = StacApiIO(conformance=["https://api.stacspec.org/v1.0.0-beta.1/core"])
+
+        # Check that this does not raise an exception
+        assert conformant_io.conforms_to(ConformanceClasses.CORE)